+++ packages/compartment-mapper/src/types.js
@@ -33,8 +33,8 @@
  * @property {string} location
  * @property {Record<string, ModuleDescriptor>} modules
  * @property {Record<string, ScopeDescriptor>} scopes
- * @property {Record<string, ParserDescriptor>} parsers - type for extension
- * @property {Record<string, ParserDescriptor>} types - type for module specifier
+ * @property {Record<string, Language>} parsers - language for extension
+ * @property {Record<string, Language>} types - language for module specifier
  */
 
 /**
@@ -45,7 +45,7 @@
  * @property {string} [compartment]
  * @property {string} [module]
  * @property {string} [location]
- * @property {ParserDescriptor} [parser]
+ * @property {Language} [parser]
  * @property {string} [exit]
  */
 
@@ -62,7 +62,7 @@
  */
 
 /**
- * @typedef {'mjs' | 'cjs' | 'json' | 'pre'} ParserDescriptor
+ * @typedef {'mjs' | 'cjs' | 'json' | 'premjs' | 'precjs'} Language
  */
 
 // /////////////////////////////////////////////////////////////////////////////
@@ -128,7 +128,7 @@
  * @param {string} packageLocation
  * @returns {Promise<{
  *   bytes: Uint8Array,
- *   parser: ParserDescriptor,
+ *   parser: Language,
  *   record: FinalStaticModuleType,
  * }>}
  */
@@ -144,8 +144,13 @@
  */
 
 /**
- * @typedef {ExecuteOptions & Object} AssemblyOptions
+ * @typedef {Record<string, ParseFn>} ParserForLanguage
+ */
+
+/**
+ * @typedef {ExecuteOptions & Object} LinkOptions
  * @property {AssembleImportHook} makeImportHook
+ * @property {ParserForLanguage} parserForLanguage
  * @property {ModuleTransforms} [moduleTransforms]
  */
 
@@ -159,7 +164,7 @@
  * @param {string} specifier
  * @param {string} location
  * @param {string} packageLocation
- * @returns {Promise<{bytes: Uint8Array, parser: ParserDescriptor}>}
+ * @returns {Promise<{bytes: Uint8Array, parser: Language}>}
  */
 
 // /////////////////////////////////////////////////////////////////////////////
@@ -179,14 +184,14 @@
  * @typedef {Object} ModuleSource
  * @property {string} [location]
  * @property {Uint8Array} [bytes]
- * @property {ParserDescriptor} [parser]
+ * @property {Language} [parser]
  * @property {string} [exit]
  */
 
 /**
  * @typedef {Object} Artifact
  * @property {Uint8Array} bytes
- * @property {ParserDescriptor} parser
+ * @property {Language} parser
  */
 
 /**
