+++ packages/compartment-mapper/test/app.agar-make.js
@@ -1,3 +1,11 @@
+// This script regenerates app.agar, an archived application generated
+// from the test fixtures.
+// This should be done *manually*, *rarely* and *deliberately*.
+// The checked-in archive exists to verify that the current archive importer
+// can recognize an archive generated with a previous version.
+// The archive may need to be regenerated if the test fixture and assertions
+// have been changed.
+
 import 'ses';
 import fs from 'fs';
 import { writeArchive } from '../src/main.js';
